Remove bucket-details in DAG
